small UI changes

+++ CODE/ZdravoCorpAppTim22/View/Manager/ViewModels/WarehouseViewModels/WarehouseViewModel.cs
@@ -8,6 +8,7 @@ using System.Linq;
 using ZdravoCorpAppTim22.Controller;
 using ZdravoCorpAppTim22.View.Manager.Commands;
 using ZdravoCorpAppTim22.View.Manager.Pages.WarehousePages;
+using ZdravoCorpAppTim22.View.Manager.Views;
 
 namespace ZdravoCorpAppTim22.View.Manager.ViewModels.WarehouseViewModels
 {
@@ -128,24 +129,27 @@ namespace ZdravoCorpAppTim22.View.Manager.ViewModels.WarehouseViewModels
         }
         public void DeleteEquipment(object obj)
         {
-            List<Equipment> selectedEquipment = ((IList)obj).Cast<Equipment>().ToList();
-            Equipment equipment = selectedEquipment[0];
-            EquipmentCollection.Remove(equipment);
+            if (ConfirmModal.Show("Are you sure?"))
+            {
+                List<Equipment> selectedEquipment = ((IList)obj).Cast<Equipment>().ToList();
+                Equipment equipment = selectedEquipment[0];
+                EquipmentCollection.Remove(equipment);
 
-            List<Equipment> eqToRemove = new List<Equipment>();
-            foreach (Equipment eq in EquipmentController.Instance.GetAll())
-            {
-                if (eq.EquipmentData.Id == equipment.EquipmentData.Id)
+                List<Equipment> eqToRemove = new List<Equipment>();
+                foreach (Equipment eq in EquipmentController.Instance.GetAll())
                 {
-                    eqToRemove.Add(eq);
+                    if (eq.EquipmentData.Id == equipment.EquipmentData.Id)
+                    {
+                        eqToRemove.Add(eq);
+                    }
                 }
+                foreach (Equipment eq in eqToRemove)
+                {
+                    eq.Room = null;
+                    EquipmentController.Instance.DeleteByID(eq.Id);
+                }
+                EquipmentDataController.Instance.DeleteByID(equipment.EquipmentData.Id);
             }
-            foreach (Equipment eq in eqToRemove)
-            {
-                eq.Room = null;
-                EquipmentController.Instance.DeleteByID(eq.Id);
-            }
-            EquipmentDataController.Instance.DeleteByID(equipment.EquipmentData.Id);
         }
         public void OpenRelocation(object obj)
         {
